+++ src/Ext/Rules/Body.h
@@ -306,7 +306,9 @@ public:
 		Valueable<bool> FallingDownTargetingFix;
 		Valueable<bool> AIAirTargetingFix;
 		Valueable<bool> OpenTopped_DecloakToFire;
+		Valueable<bool> OpenTopped_FireWhileMoving;
 		Valueable<bool> OpenTopped_AllowFiringIfAttackedByLocomotor;
+		Valueable<bool> OpenTransport_FireWhileMoving;
 
 		Valueable<bool> SortCameoByName;
 
@@ -633,7 +635,9 @@ public:
 			, FallingDownTargetingFix { false }
 			, AIAirTargetingFix { false }
 			, OpenTopped_DecloakToFire { false }
+			, OpenTopped_FireWhileMoving { true }
 			, OpenTopped_AllowFiringIfAttackedByLocomotor { true }
+			, OpenTransport_FireWhileMoving { true }
 
 			, SortCameoByName { false }