+++ src/Ext/Techno/Hooks.Transport.cpp
@@ -844,7 +844,7 @@ DEFINE_HOOK(0x519776, InfantryClass_UpdatePosition_NoQueueUpToEnter, 0x5)
 	return 0;
 }
 
-DEFINE_HOOK(0x739FA2, UnitClassClass_UpdatePosition_NoQueueUpToEnter, 0x5)
+DEFINE_HOOK(0x739FA2, UnitClass_UpdatePosition_NoQueueUpToEnter, 0x5)
 {
 	GET(UnitClass*, pThis, EBP);
 	GET(BuildingClass*, pBuilding, EBX);
@@ -885,3 +885,64 @@ DEFINE_HOOK(0x739FA2, UnitClassClass_UpdatePosition_NoQueueUpToEnter, 0x5)
 }
 
 #pragma endregion
+
+DEFINE_HOOK(0x4D9510, FootClass_SetDestination_OpenToppedFireWhileMoving, 0x6)
+{
+	GET(FootClass*, pThis, EBP);
+	GET(void*, moving, ESI);
+
+	if (!moving)
+		return 0;
+
+	auto const pUnit = abstract_cast<UnitClass*, true>(pThis);
+
+	if (!pUnit)
+		return 0;
+
+	auto const pType = pUnit->Type;
+
+	if (pType->OpenTopped && pUnit->Passengers.NumPassengers > 0)
+	{
+		const bool fireWhileMoving = TechnoTypeExt::ExtMap.Find(pType)->OpenTopped_FireWhileMoving.Get(RulesExt::Global()->OpenTopped_FireWhileMoving);
+		auto pPassenger = pUnit->Passengers.GetFirstPassenger();
+
+		while (pPassenger)
+		{
+			const bool canFire = fireWhileMoving && TechnoExt::ExtMap.Find(pPassenger)->TypeExtData->OpenTransport_FireWhileMoving.Get(RulesExt::Global()->OpenTransport_FireWhileMoving);
+
+			// Technically these 2 can't exist in the same time, but just in case
+			if (auto const pLocoTarget = pPassenger->LocomotorTarget)
+			{
+				if (!canFire)
+				{
+					pPassenger->ReleaseLocomotor(true);
+				}
+				else if (auto const pWeapon = pPassenger->GetWeapon(pPassenger->SelectWeapon(pLocoTarget))->WeaponType)
+				{
+					if (pWeapon->Warhead->IsLocomotor && !pWeapon->FireWhileMoving)
+						pPassenger->ReleaseLocomotor(true);
+				}
+			}
+
+			if (auto const pTemporal = pPassenger->TemporalImUsing)
+			{
+				if (auto const pTarget = pTemporal->Target)
+				{
+					if (!canFire)
+					{
+						pTemporal->LetGo();
+					}
+					else if (auto const pWeapon = pPassenger->GetWeapon(pPassenger->SelectWeapon(pTarget))->WeaponType)
+					{
+						if (pWeapon->Warhead->Temporal && !pWeapon->FireWhileMoving)
+							pTemporal->LetGo();
+					}
+				}
+			}
+
+			pPassenger = abstract_cast<FootClass*>(pPassenger->NextObject);
+		}
+	}
+
+	return 0;
+}